ui: Move a number of components from lib/ui-components to apps/ui/components
Change-type: patch
