+++ apps/ui/components/MermaidEditor/monaco-theme.js
