+++ apps/ui/components/StepsFlow.spec.jsx
@@ -6,7 +6,7 @@
 
 import {
 	getWrapper
-} from '../../test/ui-setup'
+} from '../../../test/ui-setup'
 import ava from 'ava'
 import {
 	mount,
