+++ apps/ui/components/UserStatusMenuItem.spec.jsx
@@ -4,7 +4,7 @@
  * Proprietary and confidential.
  */
 
-import '../../test/ui-setup'
+import '../../../test/ui-setup'
 import React from 'react'
 import ava from 'ava'
 import {
@@ -12,7 +12,7 @@ import {
 } from 'enzyme'
 import sinon from 'sinon'
 import UserStatusMenuItem from './UserStatusMenuItem'
-import user from '../core/cards/user'
+import user from '../../../lib/core/cards/user'
 
 const DND = {
 	title: 'Do Not Disturb',
