+++ apps/ui/lens/full/Repository/Repository.jsx
@@ -22,7 +22,7 @@ import {
 } from 'rendition'
 import styled from 'styled-components'
 import Segment from '../../common/Segment'
-import CardFields from '../../../../../lib/ui-components/shame/CardFields'
+import CardFields from '../../../components/CardFields'
 import Icon from '../../../../../lib/ui-components/shame/Icon'
 import CardLayout from '../../../layouts/CardLayout'
 import {
