+++ apps/ui/lens/full/SupportThread/index.jsx
@@ -32,7 +32,7 @@ import SlideInFlowPanel from '../../../components/Flows/SlideInFlowPanel'
 import * as helpers from '../../../../../lib/ui-components/services/helpers'
 import Timeline from '../../list/Timeline'
 import CardLayout from '../../../layouts/CardLayout'
-import CardFields from '../../../../../lib/ui-components/shame/CardFields'
+import CardFields from '../../../components/CardFields'
 import Event from '../../../../../lib/ui-components/Event'
 import RouterLink from '../../../../../lib/ui-components/Link'
 import Collapsible from '../../../../../lib/ui-components/Collapsible'
